+++ app/src/main/java/x/x/memlists/feature/memos/MemosUiState.kt
@@ -9,6 +9,8 @@ data class MemosUiState(
     val selectedFolder: MemoFolderType? = null,
     val items: List<MemoItemSummary> = emptyList(),
     val folders: List<MemoFolderSummary> = emptyList(),
-    val selectedTags: Set<String> = emptySet()
+    val selectedTags: Set<String> = emptySet(),
+    val userFilter: UserFilter = UserFilter(),
+    val savedFolderUserFilter: UserFilter? = null
 )